# openapi.yaml
openapi: 3.0.1 
info:
    title: AI Sportsbook
    description: Retrieve current odds for various sports and historical data for sports teams and players.
    version: "v1"
servers:
    - url: https://ai-sportsbook.replit.app/
paths:
...
    /sports/odds:
        get:
            operationId: getOdds
            summary: >-
                Returns a list of upcoming and live games with recent odds. 
                Only includes Moneyline (h2h), Spreads (spreads), Total points bets (total), and Outrights
            description: >-
                Retrieves odds for a given sport, region, and market. It includes
                upcoming and live games. Only includes Moneyline (h2h), Spreads (spreads), Total points bets (total), and Outrights. Does not
                include Player prop bets, Alternate spreads, Game period bets, or additional markets.
            parameters:
              - in: query
                name: sport
                required: true
                schema:
                    type: string
                description: The sport key obtained from calling the /sports endpoint. 'upcoming' is always valid.
                example: 'americanfootball_nfl' # Single example for path parameter
              - in: query
                name: regions
                required: true
                schema:
                    type: string
                description: >-
                    Bookmaker regions, comma delimited. Valid regions: us, us2, uk, au, eu.
                example: 'us,uk' # Single example for query parameter
              - in: query
                name: markets
                schema:
                    type: array
                    $ref: '#/components/schemas/MarketKey'
                description: >-
                    Types of odds market, comma delimited. Defaults to h2h. Valid values: h2h, spreads, totals, outrights.
              - in: query
                name: dateFormat
                schema:
                    type: string
                    enum: [unix, iso]
                description: Format of timestamps in the response. Defaults to iso.
              - in: query
                name: oddsFormat
                schema:
                    type: string
                    enum: [decimal, american]
                description: Format of odds in the response. Defaults to decimal.
              - in: query
                name: eventIds
                schema:
                    type: string
                description: Comma-separated game ids to filter the response.
              - in: query
                name: bookmakers
                schema:
                    type: string
                description: Comma-separated list of bookmakers to be returned.
              - in: query
                name: commenceTimeFrom
                schema:
                    type: string
                    format: date-time
                description: >-
                    Filter to show games commencing on/after this time. ISO 8601 format. No effect if sport is 'upcoming'.
              - in: query
                name: commenceTimeTo
                schema:
                    type: string
                    format: date-time
                description: >-
                    Filter to show games commencing on/before this time. ISO 8601 format. No effect if sport is 'upcoming'.
            responses:
                '200':
                    description: Successful response with odds data
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GamesArray'
                '400':
                    description: Bad request due to invalid parameters
                '401':
                    description: Unauthorized access due to missing or invalid apiKey
                '500':
                    description: Internal server error
    /sports/events:
        get:
            operationId: getEventOdds
            summary: Returns detailed odds for a single game, including player props, game period markets, and additional markets.
            description: >-
                Returns detailed odds for a single game, including player prop bets, game period markets, and additional markets. Accepts any available betting markets using the markets parameter. 
                Game identifier (eventId) is acquired from the /sports/odds endpoint.
            parameters:
              - in: query
                name: sport
                required: true
                schema:
                    type: string
                description: The sport key obtained from calling the /sports endpoint. 'upcoming' is always valid.
                example: 'americanfootball_nfl' # Single example for path parameter
              - in: query
                name: regions
                required: true
                schema:
                    type: string
                description: >-
                    Bookmaker regions, comma delimited. Valid regions: us, us2, uk, au, eu.
                example: 'us,uk' # Single example for query parameter
              - in: query
                name: eventId
                required: true
                schema:
                    type: string
                description: >-
                    Event ids can be found in the "id" field in the response of the main odds endpoint (see /sports/odds).
                     If the event has expired (not receiving updates due to completion or cancellation), a HTTP 404 status code will be returned.
                example: 50f069186530f68051f2a978745931b5
              - in: query
                name: markets
                schema:
                    type: array
                    $ref: '#/components/schemas/MarketKey'
                description: >-
                    (Optional) Featured betting markets, NOT USED FOR PLAYER PROP BETS, comma delimited.
              - in: query
                name: playerPropsFootball
                schema:
                    type: array
                    $ref: '#/components/schemas/PlayerPropKeyFootball'
                description: >-
                    (Optional) Market keys used in the API to identify different betting markets for NFL and NCAAF player props, comma delimited.
              - in: query
                name: playerPropsBasketball
                schema:
                    type: array
                    $ref: '#/components/schemas/PlayerPropKeyBasketball'
                description: >-
                    (Optional) Market keys used in the API to identify different betting markets for NBA, NCAAB, and WNBA player propositions, comma delimited.
              - in: query
                name: playerPropsBaseball
                schema:
                    type: array
                    $ref: '#/components/schemas/PlayerPropKeyBaseball'
                description: >-
                    (Optional) Market keys used in the API to identify different betting markets for MLB player propositions, comma delimited.
              - in: query
                name: gamePeriodMarkets
                schema:
                    type: array
                    $ref: '#/components/schemas/GamePeriodKey'
                description: >-
                    (Optional) Market keys used in the API to identify different time-specific betting markets across various sports.
                    Covers Moneyline, spreads, over/under, and point totals for different periods of the game, comma delimited.
              - in: query
                name: dateFormat
                schema:
                    type: string
                    enum: [unix, iso]
                description: Format of timestamps in the response. Defaults to iso.
              - in: query
                name: oddsFormat
                schema:
                    type: string
                    enum: [decimal, american]
                description: Format of odds in the response. Defaults to decimal.
              - in: query
                name: bookmakers
                schema:
                    type: string
                description: Comma-separated list of bookmakers to be returned.
            responses:
                '200':
                    description: Successful response with odds data
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GamesArray'
                '400':
                    description: Bad request due to invalid parameters
                '401':
                    description: Unauthorized access due to missing or invalid apiKey
                '500':
                    description: Internal server error
    /sports/scores:
        get:
            operationId: getScores
            summary: Returns a list of upcoming, live, and recently completed games
            description: >-
                Retrieves scores for upcoming, live, and recently completed games for a given sport.
                Live and recently completed games include scores. Games from up to 3 days ago can
                be returned using the daysFrom parameter. Live scores update approximately every 30 seconds.
            parameters:
              - in: query
                name: sport
                required: true
                schema:
                    type: string
                description: The sport key obtained from calling the /sports endpoint.
              - in: query
                name: daysFrom
                schema:
                    type: integer
                    minimum: 1
                    maximum: 3
                description: >-
                    Optional - Number of days in the past to return completed games. Valid values are 1 to 3.
                    If missing, only live and upcoming games are returned.
              - in: query
                name: dateFormat
                schema:
                    type: string
                    enum: [unix, iso]
                description: >-
                    Optional - Format of timestamps in the response. Valid values are unix and iso (ISO 8601). Defaults to iso.
              - in: query
                name: eventIds
                schema:
                    type: string
                description: >-
                    Optional - Comma-separated game ids to filter the response.
            responses:
                '200':
                    description: Successful response with odds data
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GamesArray'
                '400':
                    description: Bad request due to invalid parameters
                '401':
                    description: Unauthorized access due to missing or invalid apiKey
                '500':
                    description: Internal server error
    /sports/teamdata:
        get:
            operationId: getTeamData
            summary: Returns information about a specific team
            description: >-
                Retrieves information about a specific team for a specific sport
            parameters:
              - in: query
                name: sport
                required: true
                schema:
                    type: string
                    $ref: '#/components/schemas/SportsTypes'
                description: Type of sport that the requested team falls under
                example: 'baseball'
              - in: query
                name: league
                required: true
                schema:
                    type: string
                    $ref: '#/components/schemas/LeagueTypes'
                description: League name abbreviated
                example: 'nfl'
              - in: query
                name: year
                required: true
                schema:
                    type: string
                description: The season year (format="YYYY") that you are requesting team data for.
                example: '2023'
              - in: query
                name: team
                required: true
                schema:
                    type: string
                description: Abbreviated team name, example CHI for Chicago Bears
                example: 'CHI'
            responses:
                '200':
                    description: Successful response with team data
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TeamData'
                '400':
                    description: Bad request due to invalid parameters
                '401':
                    description: Unauthorized access due to missing or invalid apiKey
                '500':
                    description: Internal server error
    /sports/teamevents:
        get:
            operationId: getTeamEvents
            summary: Returns season overview statistics for previous games in the season
            description: >-
                Retrieves statistical data about a season for a specific team for a specific sport
            parameters:
              - in: query
                name: sport
                required: true
                schema:
                    type: string
                    $ref: '#/components/schemas/SportsTypes'
                description: Type of sport that the requested team falls under
                example: 'baseball'
              - in: query
                name: league
                required: true
                schema:
                    type: string
                    $ref: '#/components/schemas/LeagueTypes'
                description: League name abbreviated
                example: 'nfl'
              - in: query
                name: year
                required: true
                schema:
                    type: string
                description: The season year (format="YYYY") that you are requesting team data for.
                example: '2023'
              - in: query
                name: team
                required: true
                schema:
                    type: string
                description: Abbreviated team name, example CHI for Chicago Bears
                example: 'CHI'
            responses:
                '200':
                    description: Successful response with team data
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TeamEvents'
                '400':
                    description: Bad request due to invalid parameters
                '401':
                    description: Unauthorized access due to missing or invalid apiKey
                '500':
                    description: Internal server error
components:
    schemas:
...
        GamesArray:
            type: array
            items:
                $ref: '#/components/schemas/Game'
        MarketKey:
            type: string
            description: >
                An enumeration of the market keys used in the API to identify different betting markets.
                Each key is associated with a specific market type.
            enum:
                - h2h
                - spreads
                - totals
                - outrights
                - h2h_lay
                - outrights_lay
            x-descriptions:
                - "Head to head, Moneyline: Bet on the winning team or player of a game (includes the draw for soccer)"
                - "Points spread, Handicap: The spreads market as featured by a bookmaker. Bet on the winning team after a points handicap has been applied to each team"
                - "Total points/goals, Over/Under: The totals market as featured by a bookmaker. Bet on the total score of the game being above or below a threshold"
                - "Outrights, Futures: Bet on a final outcome of a tournament or competition"
                - "Same as h2h: Bet against a h2h outcome. This market is only applicable to betting exchanges"
                - "Same as outrights: Bet against an outrights outcome. This market is only applicable to betting exchanges"
            example: h2h
        PlayerPropKeyFootball:
            type: string
            description: An enumeration of the market keys used in the API to identify different betting markets for NFL and NCAAF player props.
            enum:
                - player_pass_tds
                - player_pass_yds
                - player_pass_completions
                - player_pass_attempts
                - player_pass_interceptions
                - player_pass_longest_completion
                - player_rush_yds
                - player_rush_attempts
                - player_rush_longest
                - player_receptions
                - player_reception_yds
                - player_reception_longest
                - player_kicking_points
                - player_field_goals
                - player_tackles_assists
                - player_1st_td
                - player_last_td
                - player_anytime_td
            x-descriptions:
                player_pass_tds: "Pass Touchdowns (Over/Under): Betting on the over/under of a player's pass touchdowns."
                player_pass_yds: "Pass Yards (Over/Under): Betting on the over/under of a player's passing yards."
                player_pass_completions: "Pass Completions (Over/Under): Betting on the over/under of a player's pass completions."
                player_pass_attempts: "Pass Attempts (Over/Under): Betting on the over/under of a player's pass attempts."
                player_pass_interceptions: "Pass Intercepts (Over/Under): Betting on the over/under of a player's interceptions thrown."
                player_pass_longest_completion: "Pass Longest Completion (Over/Under): Betting on the over/under of the longest pass completion by a player."
                player_rush_yds: "Rush Yards (Over/Under): Betting on the over/under of a player's rushing yards."
                player_rush_attempts: "Rush Attempts (Over/Under): Betting on the over/under of a player's rush attempts."
                player_rush_longest: "Longest Rush (Over/Under): Betting on the over/under of the longest rush by a player."
                player_receptions: "Receptions (Over/Under): Betting on the over/under of a player's receptions."
                player_reception_yds: "Reception Yards (Over/Under): Betting on the over/under of a player's reception yards."
                player_reception_longest: "Longest Reception (Over/Under): Betting on the over/under of the longest reception by a player."
                player_kicking_points: "Kicking Points (Over/Under): Betting on the over/under of a player's total kicking points."
                player_field_goals: "Field Goals (Over/Under): Betting on the over/under of a player's field goals made."
                player_tackles_assists: "Tackles + Assists (Over/Under): Betting on the over/under of a player's combined tackles and assists."
                player_1st_td: "1st Touchdown Scorer (Yes/No): Betting on whether a player will score the first touchdown."
                player_last_td: "Last Touchdown Scorer (Yes/No): Betting on whether a player will score the last touchdown."
                player_anytime_td: "Anytime Touchdown Scorer (Yes/No): Betting on whether a player will score a touchdown at any time during a game."
        PlayerPropKeyBasketball:
            type: string
            description: An enumeration of the market keys used in the API to identify different betting markets for NBA, NCAAB, and WNBA player propositions.
            enum:
                - player_points
                - player_rebounds
                - player_assists
                - player_threes
                - player_blocks
                - player_steals
                - player_blocks_steals
                - player_turnovers
                - player_points_rebounds_assists
                - player_points_rebounds
                - player_points_assists
                - player_rebounds_assists
                - player_first_basket
                - player_double_double
                - player_triple_double
                - player_points_alternate
                - player_rebounds_alternate
                - player_assists_alternate
                - player_blocks_alternate
                - player_steals_alternate
                - player_threes_alternate
                - player_points_assists_alternate
                - player_points_rebounds_alternate
                - player_rebounds_assists_alternate
                - player_points_rebounds_assists_alternate
            x-descriptions:
                player_points: "Points (Over/Under): Betting on the over/under of a player's points."
                player_rebounds: "Rebounds (Over/Under): Betting on the over/under of a player's rebounds."
                player_assists: "Assists (Over/Under): Betting on the over/under of a player's assists."
                player_threes: "Threes (Over/Under): Betting on the over/under of a player's three-point field goals made."
                player_blocks: "Blocks (Over/Under): Betting on the over/under of a player's blocks."
                player_steals: "Steals (Over/Under): Betting on the over/under of a player's steals."
                player_blocks_steals: "Blocks + Steals (Over/Under): Betting on the over/under of a player's combined blocks and steals."
                player_turnovers: "Turnovers (Over/Under): Betting on the over/under of a player's turnovers."
                player_points_rebounds_assists: "Points + Rebounds + Assists (Over/Under): Betting on the over/under of a player's combined points, rebounds, and assists."
                player_points_rebounds: "Points + Rebounds (Over/Under): Betting on the over/under of a player's combined points and rebounds."
                player_points_assists: "Points + Assists (Over/Under): Betting on the over/under of a player's combined points and assists."
                player_rebounds_assists: "Rebounds + Assists (Over/Under): Betting on the over/under of a player's combined rebounds and assists."
                player_first_basket: "First Basket Scorer (Yes/No): Betting on whether a player will score the first basket."
                player_double_double: "Double Double (Yes/No): Betting on whether a player will achieve a double double."
                player_triple_double: "Triple Double (Yes/No): Betting on whether a player will achieve a triple double."
                player_points_alternate: "Alternate Points (Over/Under): Betting on an alternate line for a player's points."
                player_rebounds_alternate: "Alternate Rebounds (Over/Under): Betting on an alternate line for a player's rebounds."
                player_assists_alternate: "Alternate Assists (Over/Under): Betting on an alternate line for a player's assists."
                player_blocks_alternate: "Alternate Blocks (Over/Under): Betting on an alternate line for a player's blocks."
                player_steals_alternate: "Alternate Steals (Over/Under): Betting on an alternate line for a player's steals."
                player_threes_alternate: "Alternate Threes (Over/Under): Betting on an alternate line for a player's three-point field goals made."
                player_points_assists_alternate: "Alternate Points + Assists (Over/Under): Betting on an alternate line for a player's combined points and assists."
                player_points_rebounds_alternate: "Alternate Points + Rebounds (Over/Under): Betting on an alternate line for a player's combined points and rebounds."
                player_rebounds_assists_alternate: "Alternate Rebounds + Assists (Over/Under): Betting on an alternate line for a player's combined rebounds and assists."
                player_points_rebounds_assists_alternate: "Alternate Points + Rebounds + Assists (Over/Under): Betting on an alternate line for a player's combined points, rebounds, and assists."
        PlayerPropKeyBaseball:
            type: string
            description: An enumeration of the market keys used in the API to identify different betting markets for MLB player propositions.
            enum:
                - batter_home_runs
                - batter_hits
                - batter_total_bases
                - batter_rbis
                - batter_runs_scored
                - batter_hits_runs_rbis
                - batter_singles
                - batter_doubles
                - batter_triples
                - batter_walks
                - batter_strikeouts
                - batter_stolen_bases
                - pitcher_strikeouts
                - pitcher_record_a_win
                - pitcher_hits_allowed
                - pitcher_walks
                - pitcher_earned_runs
                - pitcher_outs
            x-descriptions:
                batter_home_runs: "Batter home runs (Over/Under): Betting on the over/under of a batter's home runs."
                batter_hits: "Batter hits (Over/Under): Betting on the over/under of a batter's hits."
                batter_total_bases: "Batter total bases (Over/Under): Betting on the over/under of a batter's total bases."
                batter_rbis: "Batter RBIs (Over/Under): Betting on the over/under of a batter's runs batted in."
                batter_runs_scored: "Batter runs scored (Over/Under): Betting on the over/under of the runs scored by a batter."
                batter_hits_runs_rbis: "Batter hits + runs + RBIs (Over/Under): Betting on the over/under of a batter's total hits, runs scored, and RBIs."
                batter_singles: "Batter singles (Over/Under): Betting on the over/under of a batter's singles."
                batter_doubles: "Batter doubles (Over/Under): Betting on the over/under of a batter's doubles."
                batter_triples: "Batter triples (Over/Under): Betting on the over/under of a batter's triples."
                batter_walks: "Batter walks (Over/Under): Betting on the over/under of a batter's base on balls."
                batter_strikeouts: "Batter strikeouts (Over/Under): Betting on the over/under of a batter's strikeouts."
                batter_stolen_bases: "Batter stolen bases (Over/Under): Betting on the over/under of a batter's stolen bases."
                pitcher_strikeouts: "Pitcher strikeouts (Over/Under): Betting on the over/under of a pitcher's strikeouts."
                pitcher_record_a_win: "Pitcher to record a win (Yes/No): Betting on whether a pitcher will record a win."
                pitcher_hits_allowed: "Pitcher hits allowed (Over/Under): Betting on the over/under of the hits allowed by a pitcher."
                pitcher_walks: "Pitcher walks (Over/Under): Betting on the over/under of the walks issued by a pitcher."
                pitcher_earned_runs: "Pitcher earned runs (Over/Under): Betting on the over/under of the earned runs allowed by a pitcher."
                pitcher_outs: "Pitcher outs (Over/Under): Betting on the over/under of the outs recorded by a pitcher."
...
        SportsTypes:
            type: string
            description: An enumeration of acceptable sport parameters
            enum:
                - football
                - baseball
                - basketball
        LeagueTypes:
            type: string
            description: An enumeration of acceptable league parameters
            enum:
                - nfl # National Football League (NFL)
                - mlb # Major League Baseball (MLB)
                - nba # National Basketball Association (NBA)
        TeamData:
            type: object
            description: >-
                General information about the team including where they are located, display name, 
                and online links to various logos and team information
            properties:
                id:
                    type: string
                    description: The identifer for the team
                location:
                    type: string
                    description: Where the team is located
                displayName:
                    type: string
                    description: The human readable display name of the team
                logos:
                    type: array
                    description: Contains online links to team logo images
                    items:
                        $ref: '#/components/schemas/Logo'
                links:
                    type: array
                    description: Contains online links to additional team information
                    items:
                        $ref: '#/components/schemas/Link'
        Logo:
            type: object
            properties:
                href:
                    type: string
                    description: The online link to the image
                width:
                    type: integer
                    description: The width of the linked image in pixels
                height:
                    type: integer
                    description: The height of the linked image in pixels
                alt:
                    type: string
                rel:
                    type: array
                    items:
                        type: string
                lastUpdated:
                    type: string
                    description: The last time the image was updated
                    format: date-time
...
        TeamEvents:
            type: object
            properties:
                games:
                    $ref: '#/components/schemas/TeamGames'
                statsMetadata:
                    $ref: '#/components/schemas/StatMetadata'